Split history.py into multiple files, one for definitions, one for classes, one for tools, and one main file
