+++ Oanda/modules/info/retrieval/__init__.py
@@ -2,5 +2,8 @@ from os.path import dirname, basename, isfile, join
 import glob
 from .history import *
 from .preprocessing import *
+from .tools import *
+from .classes import *
+from .definitions import *
 modules = glob.glob(join(dirname(__file__), "*.py"))
 __all__ = [ basename(f)[:-3] for f in modules if isfile(f) and not f.startswith('_')]
